Added Instruction Screen + button to start playing
With a pretty zelda fairy inspired Instruction screen where game = false which changes the speed  to 0 and y position to be unseen.
when the tongue touches the start button calculated with distance, game = true and starts ( thus removing the instruction screen.
I think either added wings to the flies aswell ( or last commit)