Added fonts, placed elements better, added changing colors to warn of death
Downloaded and preloaded fonts and places the text in the space better (position).
Timed the changeSpeech better (subjective).
Made the water more gradually turn green the Weight number increases of the pollution bar.
Made the frog's health bar turn to yellow and red if the weight number of the health bar is close to 0.
Changed the numbers a bit of the what the fairies and fly do to health and pollution.
I'm sure why it's coping all the folder ?