Touched up some attributes and added sound
Match the color of of the health bar to the frog, Change the fairy's reset x position so it comes more frequently, added health to the frog when eaten (tongue overlap), lessen the speed. Added the total fly population score ( to add a possibility to have a win ! ) Added sound when User interacts/opens the interface and music when game === true.
(Tried again to add sad yoshi sound when Game Over I can't make it not looping )